requirements.txt
asyncio-mqtt==0.16.2
colorlog==6.9.0
openai==1.93.0
pip==25.1.1
py-cord==2.6.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
setuptools==65.5.0
beautifulsoup4==4.12.2
pandas==2.2.3
requests==2.31.0
aiohttp==3.9.1
playwright==1.53.0
pytz==2024.1